Moved vests to their own file
